fixed fog distance

+++ assets/shaders/star.wgsl
@@ -15,7 +15,7 @@ struct PlayerUniform {
     inv_p: mat4x4<f32>,
     origin: vec3<f32>,
     forward: vec3<f32>,
-    render_distance: f32,
+    render_distance: u32,
     znear: f32,
     zfar: f32,
 }
